# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Money Transfer App API",
    "version": "1.0.0",
    "description": "API documentation for the Money Transfer App"
  },
  "basePath": "/",
  "paths": {
...
    "/api/users/{user_id}": {
      "get": {
        "summary": "View user profile",
        "tags": ["Users"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile details"
          }
        }
      },
      "put": {
        "summary": "Update user profile",
        "tags": ["Users"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile updated successfully"
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "summary": "Get all users",
        "tags": ["Admin"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all users"
          }
        }
      },
      "post": {
        "summary": "Create a new user (Admin)",
        "tags": ["Admin"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        }
      },
      "put": {
        "summary": "Update user details (Admin)",
        "tags": ["Admin"],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "User details updated successfully"
          }
        }
      },
      "delete": {
        "summary": "Delete a user (Admin)",
        "tags": ["Admin"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {"BearerAuth": []}
  ]
}